Remove use of File::Slurp and update the dzil build.
File::Slurp can mess up some line endings when Perl layers are involved.
So, we remove the use of File::Slurp in favor of Path::Tiny and its
slurp_utf8 function. This results in a Do What I Mean solution that
shouldn't leave any unsuspected behavior.

One section that read the DATA handle using File::Slurp was changed into
the more traditional { local $/; <DATA> } type of slurping method.
However, since the need is there to read the DATA handle more than once,
we need to put the cursor back where we found it each time. This whole
process is wrapped in a try/catch block. We left the conversion of the
DATA handle to a globref in place for the time being.

We got rid of the PodWeaver stuff in favor of just writing out the
documentation by hand as it was only being used lightly anyway. During
this time, we fixed some POD errors as well as some spelling mistakes.

Updated the DZil build to be a bit more modern using the ::Starter
bundle.

Also, worked up a .mailmap to ensure all of the previous author's work
would be attributed correctly and show up with the right CPAN account
linked in metacpan.